requirements.txt
fastapi
uvicorn
pillow
imagehash